Change graph nodes to be stored centrally
Changed the Graph interface to store the nodes and instead the scheduler will get new node pointers each spin. This commit doesn't build as the scheduler has to be changed to give table names to new query node data structure and the rest of the configuring has to take this into consideration.